fix: proper error handling and input validation
- Fix silent serialization error in server.rs (unwrap_or_default -> Result)
- Fix password hashing to return proper errors instead of empty string
- Add input validation for username, email, and password
- Add ship tracking and combat methods to GameInstanceManager
- Fix clippy warnings (clone on Copy, unnecessary cast, missing Default impl)

+++ src/game/combat.rs
@@ -316,7 +316,7 @@ impl SpecialAbilityManager {
             if ability.duration_ms == 0 {
                 return false;
             }
-            let elapsed = (now - ability.start_time).num_milliseconds() as i64;
+            let elapsed = (now - ability.start_time).num_milliseconds();
             elapsed < ability.duration_ms
         });
